requirements.txt
fastapi
uvicorn
pymupdf
pdfplumber
sentence-transformers
faiss-cpu
python-multipart
sqlalchemy
psycopg2-binary
google-genai
python-dotenv
pillow
pytesseract
langgraph
